Unify Lightning Kokkos device and Lightning Qubit device under a Lightning Base device (#876)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [X] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [X] Ensure that the test suite passes, by running `make test`.

- [X] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [X] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Following the design of **Lightning Qubit* uses the New Device API end
to end, we can now migrate lightning.kokkos device to the new device
API.

**Description of the Change:**
Increase the reusability of the code by creating base classes for
Lightning StateVector, Measurements, and AjointJacobian. Create a new
device base class for LightningQubit and LightningKokkos.

**Benefits:**
Improve the maintenance and readability of the Python front-end devices
for LQ and LK.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-71783]

---------
