+++ pennylane_lightning/lightning_kokkos/lightning_kokkos.py
@@ -11,7 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
 r"""
 This module contains the :class:`~.LightningKokkos` class, a PennyLane simulator device that
 interfaces with C++ for fast linear algebra calculations.
@@ -19,13 +18,12 @@ interfaces with C++ for fast linear algebra calculations.
 import os
 import sys
 from dataclasses import replace
-from numbers import Number
 from pathlib import Path
-from typing import Callable, Optional, Sequence, Tuple, Union
+from typing import Optional
 
 import numpy as np
 import pennylane as qml
-from pennylane.devices import DefaultExecutionConfig, Device, ExecutionConfig
+from pennylane.devices import DefaultExecutionConfig, ExecutionConfig
 from pennylane.devices.default_qubit import adjoint_ops
 from pennylane.devices.modifiers import simulator_tracking, single_tape_support
 from pennylane.devices.preprocess import (
@@ -40,16 +38,21 @@ from pennylane.devices.preprocess import (
 from pennylane.measurements import MidMeasureMP
 from pennylane.operation import DecompositionUndefinedError, Operator, Tensor
 from pennylane.ops import Prod, SProd, Sum
-from pennylane.tape import QuantumScript, QuantumTape
+from pennylane.tape import QuantumScript
 from pennylane.transforms.core import TransformProgram
-from pennylane.typing import Result, ResultBatch
+from pennylane.typing import Result
+
+from pennylane_lightning.core.lightning_newAPI_base import (
+    LightningBase,
+    QuantumTape_or_Batch,
+    Result_or_ResultBatch,
+)
 
 from ._adjoint_jacobian import LightningKokkosAdjointJacobian
 from ._measurements import LightningKokkosMeasurements
 from ._state_vector import LightningKokkosStateVector
 
 try:
-    # pylint: disable=import-error, no-name-in-module
     from pennylane_lightning.lightning_kokkos_ops import backend_info, print_configuration
 
     LK_CPP_BINARY_AVAILABLE = True
@@ -57,171 +60,7 @@ except ImportError:
     LK_CPP_BINARY_AVAILABLE = False
     backend_info = None
 
-Result_or_ResultBatch = Union[Result, ResultBatch]
-QuantumTapeBatch = Sequence[QuantumTape]
-QuantumTape_or_Batch = Union[QuantumTape, QuantumTapeBatch]
-PostprocessingFn = Callable[[ResultBatch], Result_or_ResultBatch]
-
-
-def simulate(
-    circuit: QuantumScript,
-    state: LightningKokkosStateVector,
-    postselect_mode: str = None,
-) -> Result:
-    """Simulate a single quantum script.
-
-    Args:
-        circuit (QuantumTape): The single circuit to simulate
-        state (LightningKokkosStateVector): handle to Lightning state vector
-        postselect_mode (str): Configuration for handling shots with mid-circuit measurement
-            postselection. Use ``"hw-like"`` to discard invalid shots and ``"fill-shots"`` to
-            keep the same number of shots. Default is ``None``.
-
-    Returns:
-        Tuple[TensorLike]: The results of the simulation
-
-    Note that this function can return measurements for non-commuting observables simultaneously.
-    """
-    if circuit.shots and (any(isinstance(op, MidMeasureMP) for op in circuit.operations)):
-        results = []
-        aux_circ = qml.tape.QuantumScript(
-            circuit.operations,
-            circuit.measurements,
-            shots=[1],
-            trainable_params=circuit.trainable_params,
-        )
-        for _ in range(circuit.shots.total_shots):
-            state.reset_state()
-            mid_measurements = {}
-            final_state = state.get_final_state(
-                aux_circ, mid_measurements=mid_measurements, postselect_mode=postselect_mode
-            )
-            results.append(
-                LightningKokkosMeasurements(final_state).measure_final_state(
-                    aux_circ, mid_measurements=mid_measurements
-                )
-            )
-        return tuple(results)
-    state.reset_state()
-    final_state = state.get_final_state(circuit)
-    return LightningKokkosMeasurements(final_state).measure_final_state(circuit)
-
-
-def jacobian(
-    circuit: QuantumTape, state: LightningKokkosStateVector, batch_obs=False, wire_map=None
-):
-    """Compute the Jacobian for a single quantum script.
-
-    Args:
-        circuit (QuantumTape): The single circuit to simulate
-        state (LightningKokkosStateVector): handle to the Lightning state vector
-        batch_obs (bool): Determine whether we process observables in parallel when
-            computing the jacobian. This value is only relevant when the lightning
-            kokkos is built with OpenMP. Default is False.
-        wire_map (Optional[dict]): a map from wire labels to simulation indices
-
-    Returns:
-        TensorLike: The Jacobian of the quantum script
-    """
-    if wire_map is not None:
-        [circuit], _ = qml.map_wires(circuit, wire_map)
-    state.reset_state()
-    final_state = state.get_final_state(circuit)
-    return LightningKokkosAdjointJacobian(final_state, batch_obs=batch_obs).calculate_jacobian(
-        circuit
-    )
-
-
-def simulate_and_jacobian(
-    circuit: QuantumTape, state: LightningKokkosStateVector, batch_obs=False, wire_map=None
-):
-    """Simulate a single quantum script and compute its Jacobian.
-
-    Args:
-        circuit (QuantumTape): The single circuit to simulate
-        state (LightningKokkosStateVector): handle to the Lightning state vector
-        batch_obs (bool): Determine whether we process observables in parallel when
-            computing the jacobian. This value is only relevant when the lightning
-            kokkos is built with OpenMP. Default is False.
-        wire_map (Optional[dict]): a map from wire labels to simulation indices
-
-    Returns:
-        Tuple[TensorLike]: The results of the simulation and the calculated Jacobian
-
-    Note that this function can return measurements for non-commuting observables simultaneously.
-    """
-    if wire_map is not None:
-        [circuit], _ = qml.map_wires(circuit, wire_map)
-    res = simulate(circuit, state)
-    jac = LightningKokkosAdjointJacobian(state, batch_obs=batch_obs).calculate_jacobian(circuit)
-    return res, jac
-
-
-def vjp(
-    circuit: QuantumTape,
-    cotangents: Tuple[Number],
-    state: LightningKokkosStateVector,
-    batch_obs=False,
-    wire_map=None,
-):
-    """Compute the Vector-Jacobian Product (VJP) for a single quantum script.
-    Args:
-        circuit (QuantumTape): The single circuit to simulate
-        cotangents (Tuple[Number, Tuple[Number]]): Gradient-output vector. Must
-            have shape matching the output shape of the corresponding circuit. If
-            the circuit has a single output, ``cotangents`` may be a single number,
-            not an iterable of numbers.
-        state (LightningKokkosStateVector): handle to the Lightning state vector
-        batch_obs (bool): Determine whether we process observables in parallel when
-            computing the VJP. This value is only relevant when the lightning
-            kokkos is built with OpenMP.
-        wire_map (Optional[dict]): a map from wire labels to simulation indices
-
-    Returns:
-        TensorLike: The VJP of the quantum script
-    """
-    if wire_map is not None:
-        [circuit], _ = qml.map_wires(circuit, wire_map)
-    state.reset_state()
-    final_state = state.get_final_state(circuit)
-    return LightningKokkosAdjointJacobian(final_state, batch_obs=batch_obs).calculate_vjp(
-        circuit, cotangents
-    )
-
-
-def simulate_and_vjp(
-    circuit: QuantumTape,
-    cotangents: Tuple[Number],
-    state: LightningKokkosStateVector,
-    batch_obs=False,
-    wire_map=None,
-):
-    """Simulate a single quantum script and compute its Vector-Jacobian Product (VJP).
-    Args:
-        circuit (QuantumTape): The single circuit to simulate
-        cotangents (Tuple[Number, Tuple[Number]]): Gradient-output vector. Must
-            have shape matching the output shape of the corresponding circuit. If
-            the circuit has a single output, ``cotangents`` may be a single number,
-            not an iterable of numbers.
-        state (LightningKokkosStateVector): handle to the Lightning state vector
-        batch_obs (bool): Determine whether we process observables in parallel when
-            computing the jacobian. This value is only relevant when the lightning
-            kokkos is built with OpenMP.
-        wire_map (Optional[dict]): a map from wire labels to simulation indices
-
-    Returns:
-        Tuple[TensorLike]: The results of the simulation and the calculated VJP
-    Note that this function can return measurements for non-commuting observables simultaneously.
-    """
-    if wire_map is not None:
-        [circuit], _ = qml.map_wires(circuit, wire_map)
-    res = simulate(circuit, state)
-    _vjp = LightningKokkosAdjointJacobian(state, batch_obs=batch_obs).calculate_vjp(
-        circuit, cotangents
-    )
-    return res, _vjp
-
-
+# The set of supported operations.
 _operations = frozenset(
     {
         "Identity",
@@ -284,8 +123,9 @@ _operations = frozenset(
         "C(BlockEncode)",
     }
 )
-# The set of supported operations.
+# End the set of supported operations.
 
+# The set of supported observables.
 _observables = frozenset(
     {
         "PauliX",
@@ -304,12 +144,10 @@ _observables = frozenset(
         "Exp",
     }
 )
-# The set of supported observables.
 
 
 def stopping_condition(op: Operator) -> bool:
     """A function that determines whether or not an operation is supported by ``lightning.kokkos``."""
-    # These thresholds are adapted from `lightning_base.py`
     # To avoid building matrices beyond the given thresholds.
     # This should reduce runtime overheads for larger systems.
     if isinstance(op, qml.QFT):
@@ -404,13 +242,14 @@ def _add_adjoint_transforms(program: TransformProgram) -> None:
     program.add_transform(validate_adjoint_trainable_params)
 
 
+# Kokkos specific methods
 def _kokkos_configuration():
     return print_configuration()
 
 
 @simulator_tracking
 @single_tape_support
-class LightningKokkos(Device):
+class LightningKokkos(LightningBase):
     """PennyLane Lightning Kokkos device.
 
     A device that interfaces with C++ to perform fast linear algebra calculations.
@@ -420,22 +259,19 @@ class LightningKokkos(Device):
 
     Args:
         wires (int): the number of wires to initialize the device with
-        sync (bool): immediately sync with host-sv after applying operations
         c_dtype: Datatypes for statevector representation. Must be one of
             ``np.complex64`` or ``np.complex128``.
         shots (int): How many times the circuit should be evaluated (or sampled) to estimate
             the expectation values. Defaults to ``None`` if not specified. Setting
             to ``None`` results in computing statistics like expectation values and
             variances analytically.
+        sync (bool): immediately sync with host-sv after applying operations
         kokkos_args (InitializationSettings): binding for Kokkos::InitializationSettings
             (threading parameters).
     """
 
-    # pylint: disable=too-many-instance-attributes
-
     # General device options
     _device_options = ("c_dtype", "batch_obs")
-    _new_API = True
 
     # Device specific options
     _CPP_BINARY_AVAILABLE = LK_CPP_BINARY_AVAILABLE
@@ -467,24 +303,25 @@ class LightningKokkos(Device):
             raise ImportError(
                 "Pre-compiled binaries for lightning.kokkos are not available. "
                 "To manually compile from source, follow the instructions at "
-                "https://pennylane-lightning.readthedocs.io/en/latest/installation.html."
+                "https://docs.pennylane.ai/projects/lightning/en/stable/dev/installation.html."
             )
 
-        super().__init__(wires=wires, shots=shots)
+        super().__init__(
+            wires=wires,
+            c_dtype=c_dtype,
+            shots=shots,
+            batch_obs=batch_obs,
+        )
 
-        if isinstance(wires, int):
-            self._wire_map = None  # should just use wires as is
-        else:
-            self._wire_map = {w: i for i, w in enumerate(self.wires)}
-
-        self._c_dtype = c_dtype
-        self._batch_obs = batch_obs
+        # Set the attributes to call the Lightning classes
+        self._set_lightning_classes()
 
         # Kokkos specific options
         self._kokkos_args = kokkos_args
         self._sync = sync
 
-        self._statevector = LightningKokkosStateVector(
+        # Creating the state vector
+        self._statevector = self.LightningStateVector(
             num_wires=len(self.wires), dtype=c_dtype, kokkos_args=kokkos_args, sync=sync
         )
 
@@ -496,12 +333,11 @@ class LightningKokkos(Device):
         """The name of the device."""
         return "lightning.kokkos"
 
-    @property
-    def c_dtype(self):
-        """State vector complex data type."""
-        return self._c_dtype
-
-    dtype = c_dtype
+    def _set_lightning_classes(self):
+        """Load the LightningStateVector, LightningMeasurements, LightningAdjointJacobian as class attribute"""
+        self.LightningStateVector = LightningKokkosStateVector
+        self.LightningMeasurements = LightningKokkosMeasurements
+        self.LightningAdjointJacobian = LightningKokkosAdjointJacobian
 
     def _setup_execution_config(self, config):
         """
@@ -554,6 +390,7 @@ class LightningKokkos(Device):
         program.add_transform(
             mid_circuit_measurements, device=self, mcm_config=exec_config.mcm_config
         )
+
         program.add_transform(
             decompose,
             stopping_condition=stopping_condition,
@@ -587,7 +424,7 @@ class LightningKokkos(Device):
             if self._wire_map is not None:
                 [circuit], _ = qml.map_wires(circuit, self._wire_map)
             results.append(
-                simulate(
+                self.simulate(
                     circuit,
                     self._statevector,
                     postselect_mode=execution_config.mcm_config.postselect_mode,
@@ -621,127 +458,50 @@ class LightningKokkos(Device):
             return True
         return _supports_adjoint(circuit=circuit)
 
-    def compute_derivatives(
+    def simulate(
         self,
-        circuits: QuantumTape_or_Batch,
-        execution_config: ExecutionConfig = DefaultExecutionConfig,
-    ):
-        """Calculate the jacobian of either a single or a batch of circuits on the device.
+        circuit: QuantumScript,
+        state: LightningKokkosStateVector,
+        postselect_mode: str = None,
+    ) -> Result:
+        """Simulate a single quantum script.
 
         Args:
-            circuits (Union[QuantumTape, Sequence[QuantumTape]]): the circuits to calculate derivatives for
-            execution_config (ExecutionConfig): a datastructure with all additional information required for execution
+            circuit (QuantumTape): The single circuit to simulate
+            state (LightningKokkosStateVector): handle to Lightning state vector
+            postselect_mode (str): Configuration for handling shots with mid-circuit measurement
+                postselection. Use ``"hw-like"`` to discard invalid shots and ``"fill-shots"`` to
+                keep the same number of shots. Default is ``None``.
 
         Returns:
-            Tuple: The jacobian for each trainable parameter
+            Tuple[TensorLike]: The results of the simulation
+
+        Note that this function can return measurements for non-commuting observables simultaneously.
         """
-        batch_obs = execution_config.device_options.get("batch_obs", self._batch_obs)
-
-        return tuple(
-            jacobian(circuit, self._statevector, batch_obs=batch_obs, wire_map=self._wire_map)
-            for circuit in circuits
-        )
-
-    def execute_and_compute_derivatives(
-        self,
-        circuits: QuantumTape_or_Batch,
-        execution_config: ExecutionConfig = DefaultExecutionConfig,
-    ):
-        """Compute the results and jacobians of circuits at the same time.
-
-        Args:
-            circuits (Union[QuantumTape, Sequence[QuantumTape]]): the circuits or batch of circuits
-            execution_config (ExecutionConfig): a datastructure with all additional information required for execution
-
-        Returns:
-            tuple: A numeric result of the computation and the gradient.
-        """
-        batch_obs = execution_config.device_options.get("batch_obs", self._batch_obs)
-        results = tuple(
-            simulate_and_jacobian(
-                c, self._statevector, batch_obs=batch_obs, wire_map=self._wire_map
+        if circuit.shots and (any(isinstance(op, MidMeasureMP) for op in circuit.operations)):
+            results = []
+            aux_circ = qml.tape.QuantumScript(
+                circuit.operations,
+                circuit.measurements,
+                shots=[1],
+                trainable_params=circuit.trainable_params,
             )
-            for c in circuits
-        )
-        return tuple(zip(*results))
+            for _ in range(circuit.shots.total_shots):
+                state.reset_state()
+                mid_measurements = {}
+                final_state = state.get_final_state(
+                    aux_circ, mid_measurements=mid_measurements, postselect_mode=postselect_mode
+                )
+                results.append(
+                    LightningKokkosMeasurements(final_state).measure_final_state(
+                        aux_circ, mid_measurements=mid_measurements
+                    )
+                )
+            return tuple(results)
 
-    def supports_vjp(
-        self,
-        execution_config: Optional[ExecutionConfig] = None,
-        circuit: Optional[QuantumTape] = None,
-    ) -> bool:
-        """Whether or not this device defines a custom vector jacobian product.
-        ``LightningKokkos`` supports adjoint differentiation with analytic results.
-        Args:
-            execution_config (ExecutionConfig): The configuration of the desired derivative calculation
-            circuit (QuantumTape): An optional circuit to check derivatives support for.
-        Returns:
-            Bool: Whether or not a derivative can be calculated provided the given information
-        """
-        return self.supports_derivatives(execution_config, circuit)
-
-    def compute_vjp(
-        self,
-        circuits: QuantumTape_or_Batch,
-        cotangents: Tuple[Number],
-        execution_config: ExecutionConfig = DefaultExecutionConfig,
-    ):
-        r"""The vector jacobian product used in reverse-mode differentiation. ``LightningKokkos`` uses the
-        adjoint differentiation method to compute the VJP.
-        Args:
-            circuits (Union[QuantumTape, Sequence[QuantumTape]]): the circuit or batch of circuits
-            cotangents (Tuple[Number, Tuple[Number]]): Gradient-output vector. Must have shape matching the output shape of the
-                corresponding circuit. If the circuit has a single output, ``cotangents`` may be a single number, not an iterable
-                of numbers.
-            execution_config (ExecutionConfig): a datastructure with all additional information required for execution
-        Returns:
-            tensor-like: A numeric result of computing the vector jacobian product
-        **Definition of vjp:**
-        If we have a function with jacobian:
-        .. math::
-            \vec{y} = f(\vec{x}) \qquad J_{i,j} = \frac{\partial y_i}{\partial x_j}
-        The vector jacobian product is the inner product of the derivatives of the output ``y`` with the
-        Jacobian matrix. The derivatives of the output vector are sometimes called the **cotangents**.
-        .. math::
-            \text{d}x_i = \Sigma_{i} \text{d}y_i J_{i,j}
-        **Shape of cotangents:**
-        The value provided to ``cotangents`` should match the output of :meth:`~.execute`. For computing the full Jacobian,
-        the cotangents can be batched to vectorize the computation. In this case, the cotangents can have the following
-        shapes. ``batch_size`` below refers to the number of entries in the Jacobian:
-        * For a state measurement, the cotangents must have shape ``(batch_size, 2 ** n_wires)``
-        * For ``n`` expectation values, the cotangents must have shape ``(n, batch_size)``. If ``n = 1``,
-          then the shape must be ``(batch_size,)``.
-        """
-        batch_obs = execution_config.device_options.get("batch_obs", self._batch_obs)
-        return tuple(
-            vjp(circuit, cots, self._statevector, batch_obs=batch_obs, wire_map=self._wire_map)
-            for circuit, cots in zip(circuits, cotangents)
-        )
-
-    def execute_and_compute_vjp(
-        self,
-        circuits: QuantumTape_or_Batch,
-        cotangents: Tuple[Number],
-        execution_config: ExecutionConfig = DefaultExecutionConfig,
-    ):
-        """Calculate both the results and the vector jacobian product used in reverse-mode differentiation.
-        Args:
-            circuits (Union[QuantumTape, Sequence[QuantumTape]]): the circuit or batch of circuits to be executed
-            cotangents (Tuple[Number, Tuple[Number]]): Gradient-output vector. Must have shape matching the output shape of the
-                corresponding circuit. If the circuit has a single output, ``cotangents`` may be a single number, not an iterable
-                of numbers.
-            execution_config (ExecutionConfig): a datastructure with all additional information required for execution
-        Returns:
-            Tuple, Tuple: the result of executing the scripts and the numeric result of computing the vector jacobian product
-        """
-        batch_obs = execution_config.device_options.get("batch_obs", self._batch_obs)
-        results = tuple(
-            simulate_and_vjp(
-                circuit, cots, self._statevector, batch_obs=batch_obs, wire_map=self._wire_map
-            )
-            for circuit, cots in zip(circuits, cotangents)
-        )
-        return tuple(zip(*results))
+        state.reset_state()
+        final_state = state.get_final_state(circuit)
+        return LightningKokkosMeasurements(final_state).measure_final_state(circuit)
 
     @staticmethod
     def get_c_interface():
