+++ pennylane_lightning/lightning_qubit/_state_vector.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """
-Class implementation for state-vector manipulation.
+Class implementation for lightning_qubit state-vector manipulation.
 """
 
 try:
@@ -26,18 +26,20 @@ except ImportError:
 
 import numpy as np
 import pennylane as qml
-from pennylane import BasisState, DeviceError, StatePrep
 from pennylane.measurements import MidMeasureMP
 from pennylane.ops import Conditional
 from pennylane.ops.op_math import Adjoint
 from pennylane.tape import QuantumScript
 from pennylane.wires import Wires
 
+# pylint: disable=ungrouped-imports
+from pennylane_lightning.core._state_vector_base import LightningBaseStateVector
+
 from ._measurements import LightningMeasurements
 
 
-class LightningStateVector:
-    """Lightning state-vector class.
+class LightningStateVector(LightningBaseStateVector):  # pylint: disable=too-few-public-methods
+    """Lightning Qubit state-vector class.
 
     Interfaces with C++ python binding methods for state-vector manipulation.
 
@@ -48,44 +50,14 @@ class LightningStateVector:
         device_name(string): state vector device name. Options: ["lightning.qubit"]
     """
 
-    def __init__(self, num_wires, dtype=np.complex128, device_name="lightning.qubit"):
-        self._num_wires = num_wires
-        self._wires = Wires(range(num_wires))
-        self._dtype = dtype
+    def __init__(self, num_wires, dtype=np.complex128):
 
-        if dtype not in [np.complex64, np.complex128]:  # pragma: no cover
-            raise TypeError(f"Unsupported complex type: {dtype}")
+        super().__init__(num_wires, dtype)
 
-        if device_name != "lightning.qubit":
-            raise DeviceError(f'The device name "{device_name}" is not a valid option.')
+        self._device_name = "lightning.qubit"
 
-        self._device_name = device_name
+        # Initialize the state vector
         self._qubit_state = self._state_dtype()(self._num_wires)
-
-    @property
-    def dtype(self):
-        """Returns the state vector data type."""
-        return self._dtype
-
-    @property
-    def device_name(self):
-        """Returns the state vector device name."""
-        return self._device_name
-
-    @property
-    def wires(self):
-        """All wires that can be addressed on this device"""
-        return self._wires
-
-    @property
-    def num_wires(self):
-        """Number of wires addressed on this device"""
-        return self._num_wires
-
-    @property
-    def state_vector(self):
-        """Returns a handle to the state vector."""
-        return self._qubit_state
 
     @property
     def state(self):
@@ -109,11 +81,6 @@ class LightningStateVector:
         """
         return StateVectorC128 if self.dtype == np.complex128 else StateVectorC64
 
-    def reset_state(self):
-        """Reset the device's state"""
-        # init the state vector to |00..0>
-        self._qubit_state.resetStateVector()
-
     def _apply_state_vector(self, state, device_wires: Wires):
         """Initialize the internal state vector in a specified state.
         Args:
@@ -135,25 +102,6 @@ class LightningStateVector:
             return
 
         self._qubit_state.setStateVector(state, list(device_wires))
-
-    def _apply_basis_state(self, state, wires):
-        """Initialize the state vector in a specified computational basis state.
-
-        Args:
-            state (array[int]): computational basis state of shape ``(wires,)``
-                consisting of 0s and 1s.
-            wires (Wires): wires that the provided computational state should be
-                initialized on
-
-        Note: This function does not support broadcasted inputs yet.
-        """
-        if not set(state.tolist()).issubset({0, 1}):
-            raise ValueError("BasisState parameter must consist of 0 or 1 integers.")
-
-        if len(state) != len(wires):
-            raise ValueError("BasisState parameter and wires must be of equal length.")
-
-        self._qubit_state.setBasisState(list(state), list(wires))
 
     def _apply_lightning_controlled(self, operation):
         """Apply an arbitrary controlled operation to the state tensor.
@@ -253,7 +201,9 @@ class LightningStateVector:
                 )
             elif isinstance(operation, qml.PauliRot):
                 method = getattr(state, "applyPauliRot")
-                paulis = operation._hyperparameters["pauli_word"]
+                paulis = operation._hyperparameters[  # pylint: disable=protected-access
+                    "pauli_word"
+                ]
                 wires = [i for i, w in zip(wires, paulis) if w != "I"]
                 word = "".join(p for p in paulis if p != "I")
                 method(wires, invert_param, operation.parameters, word)
@@ -271,47 +221,3 @@ class LightningStateVector:
                 except AttributeError:  # pragma: no cover
                     # To support older versions of PL
                     method(operation.matrix, wires, False)
-
-    def apply_operations(
-        self, operations, mid_measurements: dict = None, postselect_mode: str = None
-    ):
-        """Applies operations to the state vector."""
-        # State preparation is currently done in Python
-        if operations:  # make sure operations[0] exists
-            if isinstance(operations[0], StatePrep):
-                self._apply_state_vector(operations[0].parameters[0].copy(), operations[0].wires)
-                operations = operations[1:]
-            elif isinstance(operations[0], BasisState):
-                self._apply_basis_state(operations[0].parameters[0], operations[0].wires)
-                operations = operations[1:]
-        self._apply_lightning(
-            operations, mid_measurements=mid_measurements, postselect_mode=postselect_mode
-        )
-
-    def get_final_state(
-        self,
-        circuit: QuantumScript,
-        mid_measurements: dict = None,
-        postselect_mode: str = None,
-    ):
-        """
-        Get the final state that results from executing the given quantum script.
-
-        This is an internal function that will be called by the successor to ``lightning.qubit``.
-
-        Args:
-            circuit (QuantumScript): The single circuit to simulate
-            mid_measurements (None, dict): Dictionary of mid-circuit measurements
-            postselect_mode (str): Configuration for handling shots with mid-circuit measurement
-                postselection. Use ``"hw-like"`` to discard invalid shots and ``"fill-shots"`` to
-                keep the same number of shots. Default is ``None``.
-
-        Returns:
-            LightningStateVector: Lightning final state class.
-
-        """
-        self.apply_operations(
-            circuit.operations, mid_measurements=mid_measurements, postselect_mode=postselect_mode
-        )
-
-        return self
